Created a map package, and moved CampusSwitch into location package
