+++ app/src/main/java/com/example/campusguide/map/GoogleMapAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.campusguide
+package com.example.campusguide.map
 
 import com.google.android.gms.maps.CameraUpdate
 import com.google.android.gms.maps.CameraUpdateFactory
