+++ app/src/main/java/com/example/campusguide/search/Search.kt
@@ -11,7 +11,7 @@ import androidx.core.content.ContextCompat
 import androidx.fragment.app.FragmentActivity
 import com.example.campusguide.ActivityResultListener
 import com.example.campusguide.Constants
-import com.example.campusguide.Map
+import com.example.campusguide.map.Map
 import com.google.android.gms.common.ConnectionResult
 import com.google.android.gms.common.api.GoogleApiClient
 import com.google.android.gms.common.api.GoogleApiClient.ConnectionCallbacks
